9주차 과제(array)

+++ Hw2/hw2.c
@@ -3,12 +3,12 @@
 
 int main()
 {
-	float input;
+	double input;
 
 	printf("Please enter kilometers:");
-	scanf_s("%f", &input);
-	float output = input / 1.609;
+	scanf_s("%lf", &input);
+	double output = input / 1.609;
 	printf("%.1fkm is equal to %.1f miles\n", input, output);
-
+	
 	return 0;
 }
